Making deploy of .well-known more generic

+++ tools/deploy.sh
@@ -25,9 +25,15 @@ init() {
 
 backup() {
   mv _site/* "$_backup_dir"
-  mv _site/.well-known "$_backup_dir"
   mv .git "$_backup_dir"
-  echo "include: [\".well-known\"]" > "$_backup_dir/_config.yml"
+  # Include .well-known directory if exists
+  if [[ -f "_site/.well-known" ]]; then
+    mv _site/.well-known "$_backup_dir"
+    # Include .well-known directory in case we're deploying to Github Pages
+    if [[ -f CNAME ]]; then
+      echo "include: [\".well-known\"]" > "$_backup_dir/_config.yml"
+    fi
+  fi
 
   # When adding custom domain from Github website,
   # the CANME only exist on `gh-pages` branch
